🚚 api: move delete route to src directory
