+++ api/src/certification/session/application/session-route.js
@@ -53,6 +53,29 @@ const register = async function (server) {
         tags: ['api', 'session'],
       },
     },
+    {
+      method: 'DELETE',
+      path: '/api/sessions/{id}',
+      config: {
+        validate: {
+          params: Joi.object({
+            id: identifiersType.sessionId,
+          }),
+        },
+        pre: [
+          {
+            method: authorization.verifySessionAuthorization,
+            assign: 'authorizationCheck',
+          },
+        ],
+        handler: sessionController.remove,
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès au centre de certification**\n" +
+            "- Supprime la session et les candidats si la session n'a pas démarrée",
+        ],
+        tags: ['api', 'session'],
+      },
+    },
   ]);
 };
